Suppress CA1034
