+++ src/HomeInventory/HomeInventory.Tests/Domain/Primitives/PublicTestSubject.cs
@@ -1,0 +1,14 @@
+﻿namespace HomeInventory.Tests.Domain.Primitives;
+
+public sealed class PublicTestSubject
+{
+    public PublicTestSubject()
+    {
+    }
+
+#pragma warning disable IDE0060 // Remove unused parameter
+    public PublicTestSubject(object arg)
+#pragma warning restore IDE0060 // Remove unused parameter
+    {
+    }
+}
