+++ src/HomeInventory/HomeInventory.Tests/Systems/Mapping/GuidIdConverterTests.cs
@@ -56,7 +56,9 @@ public class GuidIdConverterTests : BaseTest<GuidIdConverterTests.GivenTestConte
     protected override GivenTestContext CreateGiven(VariablesContainer variables) =>
         new(variables, Fixture);
 
+#pragma warning disable CA1034 // Nested types should not be visible
     public sealed class GivenTestContext : GivenContext<GivenTestContext>
+#pragma warning restore CA1034 // Nested types should not be visible
     {
         public GivenTestContext(VariablesContainer variables, IFixture fixture)
             : base(variables, fixture)
